fixed some compiler warnings: double-promotion

+++ KappaAnalysis/src/Producers/LuminosityWeightProducer.cc
@@ -10,6 +10,6 @@ void LuminosityWeightProducer::Produce(KappaEvent const& event,
 		KappaProduct& product,
 		KappaSettings const& settings) const
 {
-	product.m_weights["luminosityWeight"] = (1.0 / settings.GetIntLuminosity());
+	product.m_weights["luminosityWeight"] = (1.0 / static_cast<double>(settings.GetIntLuminosity()));
 }
